requirements.txt
pandas
numpy
faker
yfinance
requests
psycopg2-binary 
python-dotenv
plotly
scikit-learn
sqlalchemy